refactoring word import
