+++ org.mwc.debrief.core/src/org/mwc/debrief/core/loaders/MsDocLoader.java
@@ -3,6 +3,7 @@ package org.mwc.debrief.core.loaders;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.InvocationTargetException;
+import java.util.ArrayList;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.Status;
@@ -49,8 +50,9 @@ public class MsDocLoader extends IPlotLoader.BaseLoader
               // ok, get reading
               if (fileName.endsWith(".doc"))
               {
-                ImportNarrativeDocument iw = new ImportNarrativeDocument(theLayers);
-                iw.importThisWord(fileName, inputStream);
+                ImportNarrativeDocument iw = new ImportNarrativeDocument(theLayers);                
+                ArrayList<String> strings = iw.importFromWord(fileName, inputStream);
+                iw.processThese(strings);
               }
 
               // and inform the plot editor
